Fix transmission.
The game doesn't know garrison missions exist in MoP, so reimplement
the code that overrides that. Taken from Bunny's backport.
Could have removed the garrison part entirely, but left it in anyway.

Was having issues with server name being included, so now it will be
removed when sender/receiver are on the same server.